map logic in mapcomponent, will likely render in getting started, need to refactor for props passing and useeffect logix
