+++ client/src/App.tsx
@@ -1,6 +1,5 @@
 import './App.css';
-import { BrowserRouter as Router, Routes, Route}
-from 'react-router-dom';
+import { BrowserRouter as Router, Routes, Route} from 'react-router-dom';
 import Home from './Pages/Home';
 import { useSession } from './Components/users/Session';
 import NavBar from './Components/home/NavBar';
